Add some reusable functions to generic test class

+++ src/Cache/Results.php
@@ -44,7 +44,11 @@ class Results extends Query {
 	 * @return mixed|array|object|null  The response or null if not found in cache
 	 */
 	public function get_query_results_from_cache_cb( $result, $request ) {
-		if ( ! Settings::caching_enabled() ) {
+
+		// if caching is not enabled or the request is authenticated, bail early
+		// right now we're not supporting GraphQL cache for authenticated requests.
+		// Possibly in the future.
+		if ( ! Settings::caching_enabled() || is_user_logged_in() ) {
 			return $result;
 		}
 		$key = $this->the_results_key( $request->params->queryId, $request->params->query, $request->params->variables, $request->params->operation );
@@ -75,7 +79,10 @@ class Results extends Query {
 		$variables,
 		$request
 	) {
-		if ( ! Settings::caching_enabled() ) {
+		// if caching is not enabled or the request is authenticated, bail early
+		// right now we're not supporting GraphQL cache for authenticated requests.
+		// Possibly in the future.
+		if ( ! Settings::caching_enabled() || is_user_logged_in() ) {
 			return;
 		}
 
@@ -113,13 +120,13 @@ class Results extends Query {
 	 * Related to the data type that changed.
 	 */
 	public function purge_nodes_cb( $type, $id, $nodes ) {
-		if ( is_array( $nodes ) ) {
+		if ( is_array( $nodes ) && ! empty( $nodes ) ) {
 			foreach ( $nodes as $request_key ) {
 				$this->delete( $request_key );
 			}
 
 			//phpcs:ignore WordPress.PHP.DevelopmentFunctions.error_log_print_r
-			graphql_debug( 'Graphql delete nodes ' . print_r( $nodes, 1 ) );
+			graphql_debug( 'Graphql delete nodes', [ 'nodes' => $nodes ] );
 		}
 	}
 }
